src/producer/requirements.txt
websockets>=11.0
boto3>=1.34.0
pyyaml>=6.0
pytest-asyncio>=0.23.0
